Clarify why header-right is being unregistered
I missed Bill's commit a few months ago that added the functionality to unregister header-right. When I installed Garys plugin, I had no idea why it vanished.  I just wanted to make it a little bit more obvious in this file what was going on, by adding a comment for it, since this was previously stuck under the "Remove unused secondary sidebar" comment